Add 5.625x internal resolution option for pixel-perfect 1080p output
192 (DS screen height) x 5.625 = 1080, so this scale renders the 3D
layer at exactly the physical pixel height of a 1080p display in
single-screen layouts, avoiding an extra resize blur step.

The internal resolution scale is threaded through as a Float end to
end (settings, JNI, RendererConfiguration) instead of an Int, since
the renderer previously assumed an integer supersampling factor. The
matching native changes are in the melonDS-android-lib submodule,
now pointed at a personal copy since I don't have push access to
rafaelvcaetano/melonDS-android-lib.

+++ app/src/main/cpp/Configuration.h
@@ -19,12 +19,12 @@ struct SoftwareRenderSettings : public RenderSettings
 struct OpenGlRenderSettings : public RenderSettings
 {
     bool betterPolygons;
-    int scale;
+    float scale;
 };
 
 struct ComputeRenderSettings : public RenderSettings
 {
-    int scale;
+    float scale;
     bool highResCoordinates;
 };
