+++ app/src/main/java/me/magnum/melonds/ui/emulator/model/RuntimeRendererConfiguration.kt
@@ -5,6 +5,6 @@ import me.magnum.melonds.domain.model.render.RenderStrategy
 
 data class RuntimeRendererConfiguration(
     val videoFiltering: VideoFiltering,
-    val resolutionScaling: Int,
+    val resolutionScaling: Float,
     val renderStrategy: RenderStrategy,
 )
